src/pinecone_setup.py
import os
from dotenv import load_dotenv
from pinecone import Pinecone, ServerlessSpec

load_dotenv()

pc = Pinecone(api_key=os.getenv("PINECONE_API_KEY"))
index_name = os.getenv("PINECONE_INDEX_NAME")
diemnsion = 1534

if index_name not in pc.list_indexes().names():
    print(f"Creating Index  {index_name}...")
    pc.create_index(
        name = index_name,
        dimension=diemnsion,
        metric="cosine",
        spec=ServerlessSpec(cloud="aws",region="us-east-1")
    )
    print("Index created !!")
else:
    print("Index already exists !!")