requirements.txt
langchain
langchain-openai
langchain-google-genai
langchain-pinecone
pinecone-client
chainlit
python-dotenv
pydantic
openai
google-generativeai